main/cpfs_cli.cpp: Always invoke bash when finding meta servers
A pipe command is used to find the metadata servers, reusing the system
configuration.  This has been done with ". <configfile> && echo ...".
But this uses /bin/sh, and is not guaranteed to have the same effect as
our init scripts which uses bash.  So we now explicitly perform the
operations using bash.

Change-Id: I0fac6273b521ee73ce6924484f61f4d563dbdae1
Signed-off-by: ClusterTech CPFS Team <cpfs@clustertech.com>