connector_impl.cpp: Retry if authenticator shuts down connection
In case the authenticator times out (common when testing), the old code
would have the retry sequence interrupted, because the connecting socket
is completed and have created a FIM socket, while the new FIM socket has
nothing to continue trying to connect. We now add a cleanup routine
active only during the authentication phase, so that if the
authenticator ever shuts down the socket, the connect action will retry.

Change-Id: I0ab330f4dd38de1504a91017fd53cffcc1655e5e
Signed-off-by: ClusterTech CPFS Team <cpfs@clustertech.com>