server/ds/resync_impl.cpp: Unify resync sender flow
Previously the optimized and non-optimized resync procedures are
quite different in the way they get the inodes for resync.  The
optimized procedure would get a list of inodes, send it to the receiver,
wait for the receiver to send an inode list back, and send the inodes
data one by one.  The non-optimized procedure would instead simply send
the inodes data.

In this commit, we removed the non-optimized flow, so that even if
non-optimized the sender will send and receive the inode list.  This is
necessary because eventually the receiver is going to orchestrate the
whole resync processes.  It thus must learn a list of inodes that
requires resync.

Change-Id: I7e207b40e17e5e5b0e84c44173ca4a6588881983
Signed-off-by: Isaac To <kkto@clustertech.com>
Reviewed-on: https://code.clustertech.com/28560