ds/resync.hpp: Partition DS resync to phases, a few inodes a time
This partition the DS resync process so that in each phase a small
number (by default 32) of inodes are resync'ed.  This is bad for total
amount of time needed for resync, but we will soon add necessary changes
so that we allow the DS to be operational during the process (except for
the few inodes that are in the process of resync), which would allow the
resync process to be online.

Change-Id: I1466bbcafc890cc79b5ecfdc456e2b389c2f6f6b
Signed-off-by: Isaac To <kkto@clustertech.com>
Reviewed-on: https://code.clustertech.com/28563