ds/worker_impl.cpp: Use new strategy to handle defer triggering/reset
Previously we make use of the state change ID.  But once we allow inodes
to be removed, that same strategy will break.

Instead, we retrieve the current state upon DeferResetFim's.  The worker
state is updated to that of the DSG state information of the DS, and all
deferred work are performed, expecting that some of them may be queued
back to the FimDeferMgr and callback slots again.  Between
DeferResetFim's, inodes that become deferred will stay deferred.  This
way, we do not even need to check the state change ID, and the Fims
of the same inode cannot be reordered.

Change-Id: I5d4505fd38c9c322556398b5d66a3e451be4b693
Signed-off-by: Isaac To <kkto@clustertech.com>
Reviewed-on: https://code.clustertech.com/28571