ds/worker_impl.cpp: Add handling of a new DS state DSGResync
Enhance the ResyncDeferment structure so that it also keeps the inodes
currently being deferred.  This is at present not really used, because
it requires somebody to set the state.  This is to be added in upcoming
commits.

Also remove the state change ID in the DeferResetFim, as it is no longer
needed.

Change-Id: I137114ff5e0c26fd0bb71140402b5c9868e488f1
Signed-off-by: Isaac To <kkto@clustertech.com>
Reviewed-on: https://code.clustertech.com/28572