req_completion.hpp: Allow ReqCompletionChecker to check any operations
Previously it is in the function signature that the operation to check
must be a IReqEntry, even though the implementation has no dependency on
it.  We now use a void* instead, allowing any object in the memory to
serve the purpose (as long as the same memory is not used for anything
else that is registered to the checker).

Change-Id: I04fb66ac93f2d7c6fb4ff243a7219a4f38f4974d
Signed-off-by: Isaac To <kkto@clustertech.com>
Reviewed-on: https://code.clustertech.com/28581