req_completion.hpp: Replace GetReqAckCallback by CompleteOp
Previously request completion is defined to be the completion of
req_entry request-reply cycles.  So it makes sense for request
completion checker set to provide a callback that is suitable for use by
the request entry.  Now that we have generalized request completion to
be any operation, it no longer makes sense.  We move those code to the
caller, which will use the new CompleteOp() API to clean up the request
completion checker.

Change-Id: I88139a98daa8e1fc8d6b574614e66793000cd4d8
Signed-off-by: Isaac To <kkto@clustertech.com>
Reviewed-on: https://code.clustertech.com/28582