ms/worker_impl.cpp: Remove avoidance of missing node for FreeDataFim
Since the system will automatically drop transient requests, and since
it is not harmful to unexpectedly deleting something that needs to be
deleted anyway, there is no need to check for missing DS before sending
it a FreeDataFim.  Doing it like this is much more robust, because that
makes it less likely to have race condition of the check -- which can be
incorrect when the system is in data resync state.

Change-Id: I4b6c5847c7db8a25f3cbb77a797ad175670c5394
Signed-off-by: Isaac To <kkto@clustertech.com>
Reviewed-on: https://code.clustertech.com/28596