ds/resync_impl.cpp: Use request for DSResyncPhaseInodeListFim
Previously a plain message is used.  This can lead to difficulties if
there is an ongoing truncation when an MS failover occurs during DS data
resync.  We instead use a replicating request, to ensure that truncation
is really stopped (even upon failover) before we will start a new data
resync phase.

Change-Id: I3a3bb8681096c383deedde4534032307f08ec456
Signed-off-by: Isaac To <kkto@clustertech.com>
Reviewed-on: https://code.clustertech.com/28597