ms/topology_impl.cpp: Remove assumption of only one DSG state change
We previous assume that at the same time there is only one DSG state
change awaiting DSs to acknowledge in the topology manager.  This
assumption is removed in this commit.

Also ensures that once a DSG moves to a shutting down state, it will
never move to a different state.

Change-Id: Ic3168f3e02da0c1c1d091579e2c37e67151eeb07
Signed-off-by: Isaac To <kkto@clustertech.com>
Reviewed-on: https://code.clustertech.com/30111