ms/fim_processors.cpp: Fix segfault due to MS sending Fims too early
The MS send truncation requests to DSs in a thread different from the
thread that accept connection, and thus they may occur concurrently.
Previously it is possible that the DS Fim socket is set, and so a
truncation request is sent to the DS, before the initial registration
success Fim is sent to the DS.  This leads to a possibility that the DS
doesn't know that the MS has announced successful connection when the
Fim is received, leading to an empty request tracker and subsequently a
segmentation fault.

We now send the reply before setting the Fim socket, removing the
possibility.

Change-Id: I21dba05d04582cae83ff893e88e690eb2ca1f06d
Signed-off-by: Isaac To <kkto@clustertech.com>
Reviewed-on: https://code.clustertech.com/30124