ms/worker_impl.cpp: Add missing case of Resync mode to redirect truncate
During DSG data resync, if we need to truncate, do it as if the DSG is
still degraded.  If the node turns out to have resync completed, the Fim
redirected to the checksum server should produce a NotDegradedFim reply,
at which point we resend the Fim to the data server.

Change-Id: Ic3556a27d59a1945bdde6f06a04d11aa9515822a
Signed-off-by: Isaac To <kkto@clustertech.com>
Reviewed-on: https://code.clustertech.com/30125