ds/fim_processors.cpp: Update condition to send final DeferResetFim's
Previously the resets are sent when the state changed from Resync to
Ready.  Because now Fims are processed during the Resync DSG state, it
needs to be changed so that it is sent when the state changed from
Recovering to Resync.  This fixes a problem that the system may deadlock
during a resync that has no data to resync.

Change-Id: Ib52ebc0674b5f3dd531b8a56c6ad52af98d79a66
Signed-off-by: Isaac To <kkto@clustertech.com>
Reviewed-on: https://code.clustertech.com/30128