ds/worker_impl.cpp: Leave DS fs xattr alone during truncation
Previously, during checksum updates, the correct file size is set to the
file during file truncation.  This is now skipped.  This is because the
extended attribute is supposed to be replicated by the MS already, and
the DS should not need to hold an extra copy.  Skipping this xattr
setting makes the system more robust.

Change-Id: If348eb8a6df3275f5266e17dca9d3e468e05a67e
Signed-off-by: Isaac To <kkto@clustertech.com>
Reviewed-on: https://code.clustertech.com/30143