ds/worker_impl.cpp: Say the DSG is degraded for a file if it is resyncing
This is more "correct" conceptually, because a file that is resyncing
should still not be treated as usable normally.

This are potential problems when the DSs may change the inode resyncing
state at a time later than the requester of a file operation.  E.g., the
MS may think an inode has finished resyncing and hence send a request to
a DS, when the DS may not yet have time to change the inode resynch'ing
status.  Whether this is a real issue is yet to be seen, and if so it
should be handled with case-specific code.

Change-Id: I2d28fdd5a1b37b0005376705f6766aaf5a0a1701
Signed-off-by: Isaac To <kkto@clustertech.com>
Reviewed-on: https://code.clustertech.com/30144