ds/resync_impl.cpp: Update the resyncing inode list before notifying
Because the update requires a write lock of the DSG state which may take
time, it is better done before we open up requests for the inode by
telling other servers that we are done with the inode.  Doing otherwise
may result in the inode still considered to be resynch'ing when other
operations like data truncation shows up, leading to incorrect results.

Change-Id: I6fa63cbe36d69783f536a896e50c25a609baf319
Signed-off-by: Isaac To <kkto@clustertech.com>
Reviewed-on: https://code.clustertech.com/30145